+++ config.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from logger import logger
 
 
-VERSION = "1.3.3"
+VERSION = "2.1.0"
 
 DEFAULT_APP_DIR = Path("/opt/autosub-server")
 if not DEFAULT_APP_DIR.exists():